moving relation status fill bar
